refactor: Merge tricys-plus into tricys
Merge tricys-plus to achieve tricys integration and batch simulation functionality
